requirements.txt
flask
openai>=1.3.7
python-dotenv
